zebyyyyyyyyyyyyyyyyyyyyyyyyy

+++ temp2/CA/src/IM.java
@@ -2,9 +2,9 @@ import java.util.ArrayList;
 
 public class IM {
 	ArrayList<String> instructions = new ArrayList<String>();
-	public String getInstruction(int pc){
+	public String getInstruction(long pc){
 		pc = pc / 4;
-		return instructions.get(pc);
+		return instructions.get((int) pc);
 	}
 	public void addIns(String temp){
 		this.instructions.add(temp);
